Remove Hadoop compressor
Simplify compression not that Hadoop compressors are not used
