+++ lib/trino-hive-formats/src/main/java/io/trino/hive/formats/compression/Codec.java
@@ -13,22 +13,48 @@
  */
 package io.trino.hive.formats.compression;
 
+import io.airlift.compress.hadoop.HadoopStreams;
+
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 
-public interface Codec
+import static java.util.Objects.requireNonNull;
+
+public final class Codec
 {
-    OutputStream createStreamCompressor(OutputStream outputStream)
-            throws IOException;
+    private final HadoopStreams hadoopStreams;
 
-    ValueCompressor createValueCompressor();
+    Codec(HadoopStreams hadoopStreams)
+    {
+        this.hadoopStreams = requireNonNull(hadoopStreams, "hadoopStreams is null");
+    }
 
-    MemoryCompressedSliceOutput createMemoryCompressedSliceOutput(int minChunkSize, int maxChunkSize)
-            throws IOException;
+    public OutputStream createStreamCompressor(OutputStream outputStream)
+            throws IOException
+    {
+        return hadoopStreams.createOutputStream(outputStream);
+    }
 
-    InputStream createStreamDecompressor(InputStream inputStream)
-            throws IOException;
+    public ValueCompressor createValueCompressor()
+    {
+        return new ValueCompressor(hadoopStreams);
+    }
 
-    ValueDecompressor createValueDecompressor();
+    public MemoryCompressedSliceOutput createMemoryCompressedSliceOutput(int minChunkSize, int maxChunkSize)
+            throws IOException
+    {
+        return new MemoryCompressedSliceOutput(hadoopStreams, minChunkSize, maxChunkSize);
+    }
+
+    public InputStream createStreamDecompressor(InputStream inputStream)
+            throws IOException
+    {
+        return hadoopStreams.createInputStream(inputStream);
+    }
+
+    public ValueDecompressor createValueDecompressor()
+    {
+        return new ValueDecompressor(hadoopStreams);
+    }
 }
